# public/swagger/swagger.json
openapi: 3.0.0
info:
  title: 'API Documentation'
  description: 'Dokumentasi API untuk aplikasi Akunurse'
  version: 1.0.0
paths:
...
  '/form1/asesor/{asesorName}':
    get:
      tags:
        - Asesor
      summary: 'Ambil data Form 1 berdasarkan nama asesor'
      operationId: d932a5e9a554afaaeabc8c70fbc83658
      parameters:
        -
          name: asesorName
          in: path
          description: 'Nama asesor untuk filter data'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Data form_1 berhasil diambil'
          content:
            application/json:
              schema:
                properties:
                  status: { type: integer, example: 200 }
                  message: { type: string, example: 'Data form_1 berdasarkan asesor_name berhasil diambil.' }
                  data: { type: array, items: { properties: { id: { type: integer, example: 1 }, asesor_name: { type: string, example: 'John Doe' }, field_name: { type: string, example: 'Isi Field Lainnya' } }, type: object } }
                type: object
        '404':
          description: 'Data tidak ditemukan'
...
  '/get-indikator-status/{user_id}':
    get:
      tags:
        - 'Form Status'
      summary: 'List Form Status untuk membuka Indikator Pk'
      description: 'Pencarian Berdasarkan user_id'
      operationId: getFormStatusByUser
      parameters:
        -
          name: user_id
          in: path
          description: 'ID of the user'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Status data retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: 'Status data retrieved successfully.' }
                  status: { properties: { form_1_status: { type: string, example: Completed, nullable: true }, form_2_status: { type: string, example: Open, nullable: true }, form_3_status: { type: string, example: 'null', nullable: true }, form_4_status: { type: string, example: 'null', nullable: true }, form_5_status: { type: string, example: null, nullable: true }, form_6_status: { type: string, example: null, nullable: true }, form_7_status: { type: string, example: 'null', nullable: true }, form_8_status: { type: string, example: 'null', nullable: true }, form_9_status: { type: string, example: 'null', nullable: true }, form_10_status: { type: string, example: 'null', nullable: true }, form_11_status: { type: string, example: 'null', nullable: true }, form_12_status: { type: string, example: 'null', nullable: true } }, type: object }
                  status_code: { type: integer, example: 200 }
                type: object
        '404':
          description: 'No progress or status found for this user ID'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'No progress found for this user ID.' }
                  status_code: { type: integer, example: 404 }
                type: object
        '500':
          description: 'Error occurred while retrieving data'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'An error occurred while retrieving data.' }
                  error: { type: string, example: 'SQL error or exception message' }
                  status_code: { type: integer, example: 500 }
                type: object
...